List both downloadable and web certificates in user dashboard
If the web certificate setting is activated, pdf certificates were no
longer displayed in the user dashboard. In this patch, we simply swap a
couple conditions to make sure that pdf certificates are displayed even
when web certificates are activated.

This should be a temporary patch and should be removed once we stop
supporting pdf certificates and migrate existing pdf certificates to web
certificates.