Adaptive Content Block: Selectively display children based on information from external service.
This commit implements the basic infrastructure that we need for taking information from external service into account. The exact way of obtaining that information is subject to change. That is, once Domoscio gives us information about what the course data will look like on their end, we'll likely need to adjust the request(s) that we make to the Domoscio API in order to obtain pending reviews for a given student and block.