Adaptive Content Block: Adjust logic for filtering pending reviews for current block and user.
Based on information from Domoscio, the API for requesting pending reviews for a given user
will return results that specify, for each pending review, the UID of the parent unit (knowledge_node_uid)
and the UID of the review question (review_question_uid).

Note that the name of the property that specifies the UID of a review question
might need to be updated based on additional information from Domoscio.