Learner dashboard: Make sure users don't get confused by due dates that are in the past.
The Domoscio API will only return a pending review for a given unit and user if the date that its "next_review_at" property specifies is in the past.

As a result, users will never see a date on a pending revision that is in the future. But up until this commit, the UI would display the due date as "before DD/MM/YYYY". This is confusing because it suggests to a user that they are constantly missing deadlines for pending revisions.

To eliminate (or at least reduce) the potential for confusion, we now interpret the due date of a pending revision as "the date that this revision became due" (instead of "the date that this revision needs to be completed by"), and display "since DD/MM/YYYY" on the dashboard.